Ported code check_threshold from displays to extmath check_threshold is basically data operation function on a valid thresholds.
