+++ nilearn/plotting/img_plotting.py
@@ -27,13 +27,13 @@ import matplotlib.pyplot as plt
 
 from .. import _utils
 from ..image import new_img_like
-from .._utils.extmath import fast_abs_percentile
+from .._utils.extmath import fast_abs_percentile, check_threshold
 from .._utils.fixes.matplotlib_backports import (cbar_outline_get_xy,
                                                  cbar_outline_set_xy)
 from .._utils.ndimage import get_border_data
 from ..datasets import load_mni152_template
 from ..image import iter_img
-from .displays import get_slicer, get_projector, check_threshold
+from .displays import get_slicer, get_projector
 from . import cm
